requirements.txt
# 基础（milvus_lite 依赖 pkg_resources，需 setuptools<82）
setuptools>=69.0.0,<82

# 多智能体框架
agentscope==1.0.16

# LLM 调用
requests==2.32.3

# CLI 界面
rich==13.9.4

# 网络搜索（Tavily）
tavily-python>=0.5.0
